requirements.txt
streamlit
dnspython
requests
disposable_email_domains
email_validator
closeio